requirements.txt
fastapi
uvicorn
langchain
openai
chromadb
python-dotenv
pydantic
tqdm
requests
PyMuPDF
requests